plug-ins: fix a type warning.
When cross-compiling for Windows, I had this warning:
> plug-ins/file-ico/ico-load.c:221:22: warning: format '%lu' expects
> argument of type 'long unsigned int', but argument 5 has type 'long
> long unsigned int' [-Wformat=]

I guess that on some platform sizeof may be a long long uint. Just cast
the result to long uint instead (I could have done the other way around,
but I doubt the struct type IcoFileEntry would ever get bigger than long
uint max!).

+++ plug-ins/file-ico/ico-load.c
@@ -219,7 +219,7 @@ ico_read_info (FILE    *fp,
     {
       g_set_error (error, G_FILE_ERROR, 0,
                    _("Could not read '%lu' bytes"),
-                   sizeof (IcoFileEntry));
+                   (long unsigned int) sizeof (IcoFileEntry));
       g_free (entries);
       return NULL;
     }
